+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file. This
 project adheres to [Semantic Versioning](http://semver.org).
 
+## Unreleased
+
+### Added
+
+  - Application core.
+  - Application configuration.
+  - Database settings.
+  - Application tests.
+
 ## 0.2.0 - 2016-12-14
 
 ### Added